Merge

+++ src/main/java/cn/nukkit/entity/InstantEffect.java
@@ -5,11 +5,11 @@ package cn.nukkit.entity;
  * Nukkit Project
  */
 public class InstantEffect extends Effect {
-    public InstantEffect(byte id, String name, int r, int g, int b) {
+    public InstantEffect(int id, String name, int r, int g, int b) {
         super(id, name, r, g, b);
     }
 
-    public InstantEffect(byte id, String name, int r, int g, int b, boolean isBad) {
+    public InstantEffect(int id, String name, int r, int g, int b, boolean isBad) {
         super(id, name, r, g, b, isBad);
     }
 }
